refactor methods
changed #over? to utilize send instead of writing the return statements out for each individual method.
removed text from `puts` statement in #choose_again, as it was specific to one case, and the method
is utilized for two cases.
add check to #place_piece to ensure the choice made is valid